% Generated by build script
\name{eia_coal}
\alias{eia_coal}
\title{Get coal data}
\usage{
eia_coal(sub_route = "", frequency = "annual",
  data = c("value"), facets = NULL, start = NULL, end = NULL,
  length = 5000, api_key = NULL)
}
\arguments{
\item{sub_route}{Coal sub-route.}
\item{frequency}{Frequency.}
\item{data}{Data columns.}
\item{facets}{Named list of facet filters.}
\item{start}{Start period.}
\item{end}{End period.}
\item{length}{Max rows.}
\item{api_key}{EIA API key.}
}
\value{tibble of coal data}
\description{Convenience wrapper for coal data routes.}
