% Generated by build script
\name{eia_explore}
\alias{eia_explore}
\title{Explore an EIA API route}
\usage{
eia_explore(route = "", api_key = NULL)
}
\arguments{
\item{route}{API route path (e.g. "electricity", "petroleum/pri/spt").}
\item{api_key}{EIA API key.}
}
\value{list with elements: id, name, description, routes, frequencies, facets, data_fields, start_period, end_period}
\description{Returns metadata for a given route or sub-route.}
