% Generated by build script
\name{eia_international}
\alias{eia_international}
\title{Get international energy data}
\usage{
eia_international(frequency = "annual", data = c("value"),
  facets = NULL, start = NULL, end = NULL, length = 5000,
  api_key = NULL)
}
\arguments{
\item{frequency}{"monthly" or "annual".}
\item{data}{Data columns.}
\item{facets}{Named list of facets.}
\item{start}{Start period.}
\item{end}{End period.}
\item{length}{Max rows.}
\item{api_key}{EIA API key.}
}
\value{tibble of international energy data}
\description{Returns country-level production, consumption, imports, exports by energy source.}
