% Generated by build script
\name{eia_natural_gas}
\alias{eia_natural_gas}
\title{Get natural gas data}
\usage{
eia_natural_gas(sub_route = "pri/sum", frequency = "monthly",
  data = c("value"), facets = NULL, start = NULL, end = NULL,
  length = 5000, api_key = NULL)
}
\arguments{
\item{sub_route}{Natural gas sub-route.}
\item{frequency}{Frequency.}
\item{data}{Data columns.}
\item{facets}{Named list of facet filters.}
\item{start}{Start period.}
\item{end}{End period.}
\item{length}{Max rows.}
\item{api_key}{EIA API key.}
}
\value{tibble of natural gas data}
\description{Convenience wrapper for natural gas data routes.}
