% Generated by build script
\name{eia_seds}
\alias{eia_seds}
\title{Get State Energy Data System (SEDS) data}
\usage{
eia_seds(state = NULL, series_id = NULL, start = NULL, end = NULL,
  length = 5000, api_key = NULL)
}
\arguments{
\item{state}{State abbreviation(s).}
\item{series_id}{SEDS series ID(s).}
\item{start}{Start year.}
\item{end}{End year.}
\item{length}{Max rows.}
\item{api_key}{EIA API key.}
}
\value{tibble: period, state, series_id, series_description, value, unit}
\description{Returns state-level production, consumption, price, and expenditure data.}
