% Generated by build script
\name{eia_steo}
\alias{eia_steo}
\title{Get Short-Term Energy Outlook data}
\usage{
eia_steo(series_id = NULL, frequency = "monthly",
  start = NULL, end = NULL, length = 5000, api_key = NULL)
}
\arguments{
\item{series_id}{STEO series ID(s).}
\item{frequency}{Frequency.}
\item{start}{Start period.}
\item{end}{End period.}
\item{length}{Max rows.}
\item{api_key}{EIA API key.}
}
\value{tibble of STEO projection data}
\description{Returns monthly short-term (18 month) energy projections.}
